+++ src/main/java/shop/tryit/domain/member/dto/MemberFormDto.java
@@ -1,4 +1,4 @@
-package shop.tryit.web.member;
+package shop.tryit.domain.member.dto;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
