+++ src/main/java/shop/tryit/web/member/MemberController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import shop.tryit.domain.member.MemberFacade;
+import shop.tryit.domain.member.dto.MemberFormDto;
 
 @Slf4j
 @Controller
